update sideBar component
