+++ task-x/src/components/SideBarComponent/index.ts
@@ -1,0 +1,2 @@
+export * from "./SideBar";
+export * from "./SideItems";
